backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
requests==2.31.0
python-dotenv==1.0.1
yt-dlp
gunicorn==21.2.0
